+++ bucky/main.py
@@ -36,7 +36,7 @@ import bucky.carbon as carbon
 import bucky.collectd as collectd
 import bucky.metricsd as metricsd
 import bucky.statsd as statsd
-import bucky.system as system
+import bucky.systemstats as systemstats
 import bucky.dockerstats as dockerstats
 import bucky.processor as processor
 from bucky.errors import BuckyError
@@ -254,7 +254,7 @@ class Bucky(object):
         if cfg.statsd_enabled:
             stypes.append(statsd.StatsDServer)
         if cfg.system_stats_enabled:
-            stypes.append(system.SystemStatsServer)
+            stypes.append(systemstats.SystemStatsServer)
         if cfg.docker_stats_enabled:
             stypes.append(dockerstats.DockerStatsServer)
